+++ maiacore/src/maiacore/python_wrapper/py_score.cpp
@@ -16,10 +16,6 @@ void ScoreClass(const py::module& m) {
     // bindings to Score class
     py::class_<Score> cls(m, "Score");
 
-    cls.def(py::init<const std::initializer_list<std::string>&, const int>(), py::arg("partsName"),
-            py::arg("numMeasures") = 20,
-            py::call_guard<py::scoped_ostream_redirect, py::scoped_estream_redirect>());
-
     cls.def(py::init<const std::vector<std::string>&, const int>(), py::arg("partsName"),
             py::arg("numMeasures") = 20,
             py::call_guard<py::scoped_ostream_redirect, py::scoped_estream_redirect>());